+++ packages/create-app-shared/src/createBaseApp.ts
@@ -14,7 +14,7 @@ function mergeDefaultConfig(defaultConfig: AppConfig, config: AppConfig) {
   return config;
 }
 
-export default ({ loadRuntimeModules, createElement, runtimeAPI = {} }) => {
+export default ({ loadRuntimeModules, createElement, runtimeAPI = {}, runtimeValue = {} }) => {
   const createBaseApp = <T = AppConfig, P = BuildConfig, S = Context>(appConfig: T, buildConfig: P, context: S, staticConfig?: any) => {
 
     // Merge default appConfig to user appConfig
@@ -26,6 +26,12 @@ export default ({ loadRuntimeModules, createElement, runtimeAPI = {} }) => {
     Object.keys(runtimeAPI).forEach((apiKey: string) => {
       runtime.registerRuntimeAPI(apiKey, runtimeAPI[apiKey]);
     });
+
+    // Assign runtime plugin internal value
+    Object.keys(runtimeValue).forEach((key: string) => {
+      runtime.setRuntimeValue(key, runtimeValue[key]);
+    });
+
     loadRuntimeModules(runtime);
 
     // Collect app lifeCyle