+++ packages/plugin-app-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.1
+
+- [feat] set `WEBPACK_CACHE_ID` value
+- [fix] modify types template for rax
+
 ## 2.0.0
 
 - [feat] support `store.disableResetPageState`, deprecate `store.resetPageStore`